requirements.txt
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
fake-useragent==1.5.1
idna==3.6
loguru==0.7.2
marshmallow==3.21.1
mypy-extensions==1.0.0
packaging==24.0
python-dotenv==1.0.1
python-socks==2.4.4
requests==2.31.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
websocket-client==1.7.0
websockets==12.0
websockets_proxy==0.1.1
